clean up functionality for adding foods to meals
