+++ lib/event-listeners.js
@@ -3,23 +3,24 @@ import {  getFoodListMeal,
           deleteFood,
           deleteFoodFromMeals,
           populateFoodFromMeals,
-          // foodListAssigner, 
           foodListFunc,
-          foodList, 
-          tableSorter          
+          foodList,
+          tableSorter,
+          pushFoodToMeals
         } from './fetch-meals'
 
-import {getFoodList, 
-        editAndUpdateCell, 
-        editCell, 
-        updateCellAndDB, 
-        updateExistingFood, 
+import {getFoodList,
+        editAndUpdateCell,
+        editCell,
+        updateCellAndDB,
+        updateExistingFood,
         postNewFood
         } from './fetch-foods'
 
-import {  orderById, 
-          orderByAsc, 
-          orderByDesc
+  import {orderById,
+          orderByAsc,
+          orderByDesc,
+          createFoodListFromCheckbox
         } from "./shared"
 
 let counter = 0
@@ -84,7 +85,7 @@ $("#calories-header").on('click', function (event) {
   counter++
   if (counter === 1) {
     tableSorter(foodListFunc(), orderByAsc)
-  } else if (counter === 2) { 
+  } else if (counter === 2) {
      tableSorter(foodListFunc(), orderByDesc)
   } else {
     tableSorter(foodListFunc(), orderById)
@@ -92,3 +93,13 @@ $("#calories-header").on('click', function (event) {
   }
 })
 
+
+$('.add-to-meal-btn').on('click', function(event) {
+  event.preventDefault()
+  let mealId = event.currentTarget.id
+  let foods = []
+  createFoodListFromCheckbox($('.food-checkbox'), foods)
+  pushFoodToMeals(foods, mealId)
+  $('.food-checkbox').prop('checked', false)
+  $('.add-to-meal-btn').blur()
+})
