+++ lib/fetch-meals.js
@@ -78,14 +78,14 @@ const getFoodListMeal = () => {
   $('#food-list-meals').html('')
   return fetch('https://vast-retreat-17218.herokuapp.com/api/v1/foods')
   .then(response => response.json())
-  .then(myJson => foodListAssigner(myJson)) 
+  .then(myJson => foodListAssigner(myJson))
   .then(() =>
     tableSorter(foodList, orderById)
-  )    
+  )
 }
 
 const tableSorter = (foodCollection, order) => {
-  
+
   foodCollection.sort(order).forEach(function (food) {
     $('#food-list-meals').append(foodTableRowIndexHtml(food))
   })
@@ -94,15 +94,21 @@ const tableSorter = (foodCollection, order) => {
 
 
 const foodListAssigner = (myJson) => {
-  foodList = myJson 
+  foodList = myJson
 }
 
 const foodListFunc = () => {
   return foodList
 }
 
-
-
+const pushFoodToMeals = (foods, mealId) => {
+  foods.forEach(function (foodId) {
+    fetch(`https://vast-retreat-17218.herokuapp.com/api/v1/meals/${mealId}/foods/${foodId}`, {
+        method: 'POST'
+    })
+    .then(() => populateFoodFromMeals())
+  })
+}
 
 // Html for Meals page
 
@@ -154,11 +160,12 @@ module.exports = {
   deleteFood,
   findFoodsInMeals,
   deleteFoodFromMeals,
-  populateFoodFromMeals, 
-  foodList, 
+  populateFoodFromMeals,
+  foodList,
   foodListAssigner,
-  foodListFunc, 
-  tableSorter
+  foodListFunc,
+  tableSorter,
+  pushFoodToMeals
 }
 
 populateFoodFromMeals()
