+++ lib/shared.js
@@ -50,9 +50,18 @@ const orderByDesc = (a, b) => {
   }
 }
 
+const createFoodListFromCheckbox = (list, foods) => {
+  list.each(function() {
+    if ($(this).is(":checked")) {
+      foods.push($(this).attr('name'))
+    }
+  })
+}
+
 module.exports = {mealTotalCalories,
                   colorCals,
                   toatsCals,
-                  orderById, 
+                  orderById,
                   orderByAsc,
-                  orderByDesc}
+                  orderByDesc,
+                  createFoodListFromCheckbox}
